dom/tests/mochitest/integer-gemm/test_preference_enable.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=650295
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 650295 -- No objects should be visible with preference disabled</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=650295">Mozilla Bug 650295</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="text/javascript">
  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPrefEnv({
    set: [["javascript.options.wasm_moz_intgemm", false]]
  }, function() {
    /* var objects = [
      "SpeechRecognition",
      "SpeechGrammar",
      "SpeechRecognitionResult",
      "SpeechRecognitionResultList",
      "SpeechRecognitionAlternative"
    ]; */

    const gemm = WebAssembly.mozIntGemm();
     for (var i = 0; i < 1; i++) {
      is(window[gemm], undefined,
         gemm + " should be undefined with pref off");
    }

    SimpleTest.finish();
  });
</script>
</pre>
</body>
</html>
